WIP: redefine section symbols with the armds specific
It seems steer is not able to resolve or rename armlink defined
symbols, because they are not created, because they are not used
in the code (it seems armlink does not consider usage of those
symbols in steer).
Let us fix that in a straight-forward way through the defines.

Here we need steer to resolve _srodata to Load$$_rodata_bug_frames_0$$Base,
due to conflict with __start_bug_frames.

Yet we will have sections overlapping in result. That scat needs more digging.